motor current and battery voltage measurement in control module
