+++ src/src.mk
@@ -2,3 +2,4 @@
 PROJSRC  = src/main.c
 PROJSRC += src/blocking_uart_driver.c
 PROJSRC += src/motor_pwm.c
+PROJSRC += src/control.c
